requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
openai==1.12.0
flask-sqlalchemy==3.1.1
python-linkedin-v2==0.8.9
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0 